Added local tests for get card request
