+++ src/test/java/co/omise/resources/CardResourceTest.java
@@ -27,17 +27,6 @@ public class CardResourceTest extends ResourceTest {
     }
 
     @Test
-    public void testGet() throws IOException, OmiseException {
-        Card card = resource().get(CARD_ID);
-        assertRequested("GET", "/customers/" + CUSTOMER_ID + "/cards/" + CARD_ID, 200);
-
-        assertEquals("card", card.getObject());
-        assertEquals("card_test_4yq6tuucl9h4erukfl0", card.getId());
-        assertEquals("JOHN DOE", card.getName());
-        assertEquals("4242", card.getLastDigits());
-    }
-
-    @Test
     public void testUpdate() throws IOException, OmiseException {
         Card.Update update = new Card.Update()
                 .name("JOHN W. DOE");
